requirements.txt
yfinance==0.2.12
mysql-connector-python~=8.0.32
pandas==1.3.0
PyMySQL~=1.0.2
SQLAlchemy~=1.4.46
pendulum~=2.1.2